Rename VintageNetLTE -> VintageNetMobile
